pyproject.toml
[project]
name = "hyperview"
version = "0.1.0"
description = "Local-First Multimodal Data Explorer with Hyperbolic Geometry"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "lancedb>=0.1.0",
    "numpy",
    "pandas",
    "python-multipart",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=9.0.0",
    "scikit-learn>=1.0.0",
    "umap-learn>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hyperview"]
